Visualizer motivated updates. Pub/sub api, better errors.
Update Readme with better description.

Add client error code and structure the same as server errors, in
an object.

All session.js api calls return Promises, for consistency with wamp.
Subscriptions return an object, and return an array of results to
subscriptions from a server publish(), for consistency with wamp.