feat(connection): forward request+client_id to life cycle methods
Add onConnect on onClose methods to ServerProtocol and LinkProtocol

onConnect(request, client_id) -> None
onClose(client_id) -> None

They are called when a websocket connection is established or closed.  Arg
request is the aiohttp HTTP request header. client_id is an opaque string that
identifies the connection.  onC

Our use case is use an HTTP cookie in the request header to authenticate the
request.

getattr -> hasattr.