feat(exec_mode): add exec_mode to start_webserver
This adds three exec_modes to start_webserver: "main", "task", and "coroutine". The "main" mode
is the previous behavior of blocking python execution and closing the loop afterwards. The "task"
mode creates a task and returns it (does not block execution and does not close the loop). The
"coroutine" just returns the coroutine without starting it. The "task" is particularly useful for
running inside a jupyter notebook, which already has an asyncio event loop running, and for which
we do not want to block python execution.

This also adds "**kwargs" to start_webserver so that it can catch extra arguments without an error.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>