fix(license): include LICENSE file in wheel
This includes the LICENSE file in the wheel via a soft link and a MANIFEST.in file. I tried to
avoid making a soft link / copy of the LICENSE file, but it appears that the MANIFEST.in file
cannot refer to files in parent directories. This appears to work, however.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>